AbstractSCM: move process_merge_conflict to SCM (bug 1940876)
